add transaction screen ui shell for single transaction
