+++ app/src/main/AndroidManifest.xml
@@ -2,27 +2,22 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="za.co.app.budgetbee">
     <!-- To access Google+ APIs: -->
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <application
-        android:name=".BudgetBeeApplication"
         android:allowBackup="true"
         android:fullBackupContent="true"
         android:icon="@mipmap/ic_budget_bee_logo"
         android:label="@string/app_name"
+        android:name=".BudgetBeeApplication"
         android:roundIcon="@mipmap/ic_budget_bee_logo_round"
         android:supportsRtl="true"
         android:theme="@style/BudgetBeeMainTheme">
-
-        <activity android:name=".ui.transaction.AddTransactionActivity" />
-
+        <activity android:name=".ui.transaction.Transaction"></activity>
+        <activity android:name=".ui.add_transaction.AddTransactionActivity" />
         <activity
-            android:name=".ui.transactions_category.TransactionCategoryAddCategoryActivity"
             android:label="@string/title_activity_add_transaction_category"
+            android:name=".ui.transactions_category.TransactionCategoryAddCategoryActivity"
             android:theme="@style/BudgetBeeMainTheme.NoActionBar" />
-
         <activity android:name=".ui.transactions_category.TransactionCategorySelectCategoryActivity" />
-
         <activity
             android:name=".ui.landing.LandingActivity"
             android:theme="@style/BudgetBeeMainTheme">
@@ -34,5 +29,7 @@
             </intent-filter>
         </activity>
     </application>
+
+    <uses-permission android:name="android.permission.INTERNET" />
 
 </manifest>
