+++ app/src/main/java/za/co/app/budgetbee/BudgetBeeAppModule.kt
@@ -6,10 +6,10 @@ import za.co.app.budgetbee.data.model.database.BudgetBeeDatabase
 import za.co.app.budgetbee.data.model.database.BudgetBeeDoa
 import za.co.app.budgetbee.data.repository.IDatabaseRepository
 import za.co.app.budgetbee.data.repository.TransactionsRepository
+import za.co.app.budgetbee.ui.add_transaction.AddTransactionPresenter
+import za.co.app.budgetbee.ui.add_transaction.IAddTransactionMvp
 import za.co.app.budgetbee.ui.landing.ILandingMvp
 import za.co.app.budgetbee.ui.landing.LandingPresenter
-import za.co.app.budgetbee.ui.transaction.ITransactionMvp
-import za.co.app.budgetbee.ui.transaction.TransactionPresenter
 import za.co.app.budgetbee.ui.transactions_category.ITransactionCategoryMvp
 import javax.inject.Singleton
 
@@ -47,7 +47,7 @@ class BudgetBeeAppModule(val application: BudgetBeeApplication) {
 
     @Provides
     @Singleton
-    fun transactionPresenter(transactionsRepository: IDatabaseRepository): ITransactionMvp.Presenter {
-        return TransactionPresenter(transactionsRepository)
+    fun transactionPresenter(transactionsRepository: IDatabaseRepository): IAddTransactionMvp.Presenter {
+        return AddTransactionPresenter(transactionsRepository)
     }
 }
