+++ app/src/main/java/za/co/app/budgetbee/ui/add_transaction/AddTransactionFragment.kt
@@ -1,4 +1,4 @@
-package za.co.app.budgetbee.ui.transaction
+package za.co.app.budgetbee.ui.add_transaction
 
 import android.os.Bundle
 import android.view.LayoutInflater
