+++ app/src/main/java/za/co/app/budgetbee/ui/add_transaction/AddTransactionPagerAdapter.kt
@@ -1,4 +1,4 @@
-package za.co.app.budgetbee.ui.transaction
+package za.co.app.budgetbee.ui.add_transaction
 
 
 import androidx.fragment.app.Fragment
