+++ app/src/main/res/layout/activity_add_transaction.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:layout_margin="@dimen/gutterSpace"
     android:orientation="vertical"
-    tools:context=".ui.transaction.AddTransactionActivity">
+    tools:context=".ui.add_transaction.AddTransactionActivity">
 
     <com.google.android.material.textfield.TextInputLayout
         android:layout_width="match_parent"
