+++ app/src/main/res/layout/activity_transaction.xml
@@ -1,0 +1,137 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".ui.transaction.Transaction">
+
+    <include
+        layout="@layout/default_toolbar"
+        android:id="@+id/toolbar" />
+
+    <LinearLayout
+        android:id="@+id/main_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/toolbar">
+
+
+        <TextView
+            android:id="@+id/category_name_text"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/gutterSpace"
+            tools:text="Salary"
+            style="@style/text_heading6" />
+
+        <View
+            android:layout_marginEnd="@dimen/gutterSpace"
+            android:layout_marginLeft="@dimen/gutterSpace"
+            android:layout_marginRight="@dimen/gutterSpace"
+            android:layout_marginStart="@dimen/gutterSpace"
+            style="@style/divider" />
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/gutterSpace"
+            android:orientation="horizontal">
+
+            <TextView
+                android:layout_width="@dimen/transaction_text_width"
+                android:layout_height="wrap_content"
+                android:text="@string/category"
+                style="@style/text_heading6_light" />
+
+            <TextView
+                android:id="@+id/category_type_text"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="@dimen/transaction_text_margin"
+                android:layout_marginStart="@dimen/transaction_text_margin"
+                tools:text="Income"
+                style="@style/text_heading6" />
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/gutterSpace"
+            android:orientation="horizontal">
+
+            <TextView
+                android:layout_width="@dimen/transaction_text_width"
+                android:layout_height="wrap_content"
+                android:text="@string/amount"
+                style="@style/text_heading6_light" />
+
+            <TextView
+                android:id="@+id/amount_text"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="@dimen/transaction_text_margin"
+                android:layout_marginStart="@dimen/transaction_text_margin"
+                tools:text="200"
+                style="@style/text_heading6" />
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/gutterSpace"
+            android:orientation="horizontal">
+
+            <TextView
+                android:layout_width="@dimen/transaction_text_width"
+                android:layout_height="wrap_content"
+                android:text="@string/date"
+                style="@style/text_heading6_light" />
+
+            <TextView
+                android:id="@+id/date_text"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="@dimen/transaction_text_margin"
+                android:layout_marginStart="@dimen/transaction_text_margin"
+                tools:text="23 September 2020"
+                style="@style/text_heading6" />
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/gutterSpace"
+            android:orientation="horizontal">
+
+            <TextView
+                android:layout_width="@dimen/transaction_text_width"
+                android:layout_height="wrap_content"
+                android:text="@string/description"
+                style="@style/text_heading6_light" />
+
+            <TextView
+                android:id="@+id/description_text"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginLeft="@dimen/transaction_text_margin"
+                android:layout_marginStart="@dimen/transaction_text_margin"
+                tools:text="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat."
+                style="@style/text_heading6" />
+        </LinearLayout>
+    </LinearLayout>
+
+    <Button
+        android:id="@+id/edit_button"
+        android:layout_width="match_parent"
+        android:layout_height="?actionBarSize"
+        android:text="@string/edit"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="@id/main_layout"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@id/main_layout"
+        app:layout_constraintVertical_bias="1.0" />
+</androidx.constraintlayout.widget.ConstraintLayout>
