+++ app/src/main/res/layout/default_toolbar.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/colorPrimary"
+    android:orientation="horizontal">
+
+        <TextView
+            android:id="@+id/screen_title"
+            android:layout_margin="@dimen/gutterSpace"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            tools:text="Details"
+            style="@style/text_toolbar" />
+</LinearLayout>
